chore: move #501 tests into dedicated files, leaving existing suites untouched (#501)
